fix: persist description clears and share the email parser
