+++ frontend/src/features/organizations/utils/organizations.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest';
-import { parseEmailList, reconcileActiveOrgId } from './organizations';
+import { reconcileActiveOrgId } from './organizations';
 
 describe('reconcileActiveOrgId', () => {
   it('keeps an id that is still in the list', () => {
@@ -16,30 +16,5 @@ describe('reconcileActiveOrgId', () => {
 
   it('leaves an explicit no-organization choice alone', () => {
     expect(reconcileActiveOrgId([1, 2], null)).toBeNull();
-  });
-});
-
-describe('parseEmailList', () => {
-  it('splits on commas, semicolons, and whitespace', () => {
-    expect(parseEmailList('a@x.io, b@x.io;c@x.io\nd@x.io').emails).toEqual([
-      'a@x.io',
-      'b@x.io',
-      'c@x.io',
-      'd@x.io',
-    ]);
-  });
-
-  it('dedupes case-insensitively, keeping the first spelling', () => {
-    expect(parseEmailList('A@x.io a@x.io').emails).toEqual(['A@x.io']);
-  });
-
-  it('separates entries that cannot be an address', () => {
-    const { emails, invalid } = parseEmailList('ok@x.io nope, also@bad');
-    expect(emails).toEqual(['ok@x.io']);
-    expect(invalid).toEqual(['nope', 'also@bad']);
-  });
-
-  it('ignores empty input', () => {
-    expect(parseEmailList('  \n ')).toEqual({ emails: [], invalid: [] });
   });
 });
