+++ frontend/src/shared/utils/utility.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest';
 
-import { formatSliceLabel, formatWindowLabel, searchUsers } from './utility';
+import { formatSliceLabel, formatWindowLabel, parseEmailList, searchUsers } from './utility';
 
 describe('formatSliceLabel', () => {
   it('labels a weekly slice with its inclusive start and end days', () => {
@@ -97,5 +97,30 @@ describe('searchUsers', () => {
     const result = searchUsers(items, (item) => item.user, 'wag');
 
     expect(result.map((item) => item.user.display_name)).toEqual(['wagnerj', 'jwagner']);
+  });
+});
+
+describe('parseEmailList', () => {
+  it('splits on commas, semicolons, and whitespace', () => {
+    expect(parseEmailList('a@x.io, b@x.io;c@x.io\nd@x.io').emails).toEqual([
+      'a@x.io',
+      'b@x.io',
+      'c@x.io',
+      'd@x.io',
+    ]);
+  });
+
+  it('dedupes case-insensitively, keeping the first spelling', () => {
+    expect(parseEmailList('A@x.io a@x.io').emails).toEqual(['A@x.io']);
+  });
+
+  it('separates entries that cannot be an address', () => {
+    const { emails, invalid } = parseEmailList('ok@x.io nope, also@bad');
+    expect(emails).toEqual(['ok@x.io']);
+    expect(invalid).toEqual(['nope', 'also@bad']);
+  });
+
+  it('ignores empty input', () => {
+    expect(parseEmailList('  \n ')).toEqual({ emails: [], invalid: [] });
   });
 });
